requirements.txt
aiosc==0.1.4
mido==1.2.9
pymonome==0.10.1
python-rtmidi==1.4.6
